test: Extract watt/ftp test to an actual test
